py/requirements.txt
# Core Dependencies
torch>=2.0.0
numpy>=1.20.0
scipy>=1.7.0

# Network and Communication
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=3.20.0
aiohttp>=3.8.0

# GPU Monitoring
pynvml>=11.0.0
psutil>=5.9.0

# Cryptography and Security
cryptography>=3.4.0
pycryptodome>=3.15.0

# Data Processing and Analysis
pandas>=1.5.0
scikit-learn>=1.0.0

# Configuration and Logging
PyYAML>=6.0
python-dotenv>=0.20.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0

# Development Tools
black>=22.0.0
isort>=5.0.0
mypy>=0.900
flake8>=4.0.0
pylint>=2.15.0

# Documentation
Sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Monitoring and Metrics
prometheus-client>=0.15.0
statsd>=3.3.0

# Type Checking
types-PyYAML>=6.0.0
types-protobuf>=3.20.0
types-setuptools>=65.0.0

# Optional Dependencies
tensorboard>=2.10.0  # For training visualization
ray>=2.0.0  # For distributed computing support